Feat: migration home page
