+++ app/page.tsx
@@ -1,8 +1,8 @@
+'use client';
 import { useEffect, useState } from 'react';
 import { NextPage } from 'next';
 import { ContentsDataType, SheetAPIReturntype } from '@/types/inSheet';
 import useSheet from '@/hooks/useSheet';
-import home from '@/styles/home/Home.module.scss';
 import NavSection from '@/components/home/NavSection';
 import ScheduleSection from '@/components/home/ScheduleSection';
 import Loading from '@/components/common/Loading';
@@ -13,7 +13,6 @@ export interface HomePageProps {
 
 const HomePage: NextPage<HomePageProps> = ({ filter = 'scheduled' }) => {
   const { data, isLoading } = useSheet();
-
   const [contents, setContents] = useState<ContentsDataType[]>([]);
   const [total, setTotal] = useState<number>(0);
 
@@ -28,14 +27,16 @@ const HomePage: NextPage<HomePageProps> = ({ filter = 'scheduled' }) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [filter, data]);
 
-  if (isLoading) {
-    return <Loading />;
-  }
-
   return (
     <>
-      <NavSection total={total} />
-      <ScheduleSection contents={contents} />
+      {isLoading ? (
+        <Loading />
+      ) : (
+        <>
+          <NavSection total={total} />
+          <ScheduleSection contents={contents} />
+        </>
+      )}
     </>
   );
 };
