+++ components/layout/Header.tsx
@@ -1,3 +1,4 @@
+'use client';
 /* eslint-disable @next/next/no-html-link-for-pages */
 import { useRef, useEffect, useState, useMemo } from 'react';
 import Link from 'next/link';
@@ -5,11 +6,11 @@ import Sidebar from '@/components/layout/Sidebar';
 import { RxHamburgerMenu } from 'react-icons/rx';
 import { BiSearchAlt } from 'react-icons/bi';
 import header from '@/components/layout/Header.module.scss';
-import { useRouter } from 'next/router';
+import { usePathname } from 'next/navigation';
 
 const Header = () => {
   const [showSidebar, setShowSidebar] = useState(false);
-  const router = useRouter();
+  const pathname = usePathname();
   const gnbRef = useRef<HTMLElement>(null);
 
   const toggleSidebar = () => {
@@ -31,7 +32,7 @@ const Header = () => {
 
   useEffect(() => {
     setShowSidebar(() => false);
-  }, [router]);
+  }, [pathname]);
 
   useEffect(() => {
     window.addEventListener('scroll', handleScroll);
